+++ agate/table/aggregate.py
@@ -1,5 +1,7 @@
 #!/usr/bin/env python
 
+from collections import OrderedDict
+
 from agate import utils
 
 
@@ -8,22 +10,24 @@ def aggregate(self, aggregations):
     Apply one or more :class:`.Aggregation` instances to this table.
 
     :param aggregations:
-        A single :class:`.Aggregation` instance or sequence of them.
+        A single :class:`.Aggregation` instance or a sequence of tuples in the
+        format :code:`(name, aggregation)`, where each :code:`aggregation` is
+        an instance of :class:`.Aggregation`.
     :returns:
         If the input was a single :class:`Aggregation` then a single result
-        will be returned. If it was a sequence then a tuple of results will
-        be returned.
+        will be returned. If it was a sequence then an :class:`.OrderedDict` of
+        results will be returned.
     """
     if utils.issequence(aggregations):
-        results = []
+        results = OrderedDict()
 
-        for agg in aggregations:
+        for name, agg in aggregations:
             agg.validate(self)
 
-        for agg in aggregations:
-            results.append(agg.run(self))
+        for name, agg in aggregations:
+            results[name] = agg.run(self)
 
-        return tuple(results)
+        return results
     else:
         aggregations.validate(self)